test(mediaplayer): split test
