+++ mediaPlayer/tests/mediaPlayerTest.cpp
@@ -13,31 +13,25 @@ using namespace Cicada;
 #include <SDL2/SDL.h>
 #include <utils/timer.h>
 
-SDL_Window *window = nullptr;
-SDL_Renderer *mVideoRender = nullptr;
-
-static void *getView()
+static SDL_Window *getView()
 {
-    if (window == nullptr) {
-        SDL_Init(SDL_INIT_VIDEO);
-        Uint32 flags = 0;
-        flags |= SDL_WINDOW_ALLOW_HIGHDPI;
-        flags |= SDL_WINDOW_RESIZABLE;
-        SDL_CreateWindowAndRenderer(1280, 720, flags, &window, &mVideoRender);
-    }
+    SDL_Window *window = nullptr;
+    SDL_Renderer *mVideoRender = nullptr;
+    SDL_Init(SDL_INIT_VIDEO);
+    Uint32 flags = 0;
+    flags |= SDL_WINDOW_ALLOW_HIGHDPI;
+    flags |= SDL_WINDOW_RESIZABLE;
+    SDL_CreateWindowAndRenderer(1280, 720, flags, &window, &mVideoRender);
     return window;
 }
 
-static void releaseView()
+static void releaseView(SDL_Window *window)
 {
-    if (mVideoRender != nullptr) {
-        SDL_DestroyRenderer(mVideoRender);
-        mVideoRender = nullptr;
-    }
-
     if (window != nullptr) {
+        SDL_Renderer *render = SDL_GetRenderer(window);
+        if(render)
+            SDL_DestroyRenderer(render);
         SDL_DestroyWindow(window);
-        window = nullptr;
     }
 }
 
@@ -55,9 +49,10 @@ static void releaseView(){
 void test_player(const string &url, OnLoop loop, void *arg)
 {
     unique_ptr<MediaPlayer> player = unique_ptr<MediaPlayer>(new MediaPlayer());
+    SDL_Window * window = getView();
     player->SetDataSource(url.c_str());
     player->SetAutoPlay(true);
-    player->SetView(getView());
+    player->SetView(window);
     player->Prepare();
     if (loop) {
         int ret = 0;
@@ -66,5 +61,5 @@ void test_player(const string &url, OnLoop loop, void *arg)
         }
     } else
         af_msleep(10000);
-    releaseView();
+    releaseView(window);
 }
